requirements.txt
autopep8==1.7.0
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Flask==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
numpy==1.23.3
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
pycodestyle==2.9.1
requests==2.28.1
six==1.16.0
toml==0.10.2
tomli==2.0.1
torch==1.12.1
torchvision==0.13.0
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
